updated repos to use kindergarden from pypi, separate kinder-baseline… (#40)
* updated repos to use kindergarden from pypi, separate kinder-baselines instead of using subdirectories of prpl-mono

* use kindergarden[kinematic2d] instead of kindergarden due to possible import issue with 3d envs

* resolved dependency

+++ src/alphatamp/approaches/practice_makes_perfect/utils.py
@@ -56,7 +56,7 @@ def beta_from_mean_and_variance(
     assert alpha > 0
     assert _beta > 0
     rv = beta(alpha, _beta)
-    assert abs(rv.mean() - mean) < 1e-6
+    assert abs(float(rv.mean()) - mean) < 1e-6
     return rv
